refactor: LessonDto and LessonDtoDetail
