+++ src/main/java/com/linked/classbridge/dto/oneDayClass/LessonDto.java
@@ -1,9 +1,6 @@
 package com.linked.classbridge.dto.oneDayClass;
 
 import com.linked.classbridge.domain.Lesson;
-import com.linked.classbridge.domain.OneDayClass;
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import lombok.AllArgsConstructor;
@@ -19,7 +16,6 @@ public class LessonDto {
     private LocalTime startTime;
     private LocalTime endTime;
     private int participantNumber;
-    private int personal;
 
     public LessonDto(Lesson lesson) {
         lessonId = lesson.getLessonId();
@@ -27,34 +23,6 @@ public class LessonDto {
         startTime = lesson.getStartTime();
         endTime = lesson.getEndTime();
         participantNumber = lesson.getParticipantNumber();
-    }
-
-    public LessonDto(Lesson lesson, int personal) {
-        lessonId = lesson.getLessonId();
-        lessonDate = lesson.getLessonDate();
-        startTime = lesson.getStartTime();
-        endTime = lesson.getEndTime();
-        participantNumber = lesson.getParticipantNumber();
-        this.personal = personal;
-    }
-
-    public record Request(
-            @Schema(description = "레슨 날짜", example = "2024-06-30")
-            @NotNull
-            LocalDate lessonDate,
-            @Schema(description = "레슨 시간", example = "18:00:00")
-            @NotNull
-            LocalTime startTime
-    ) {
-        public Lesson toEntity(OneDayClass oneDayClass) {
-            return Lesson.builder()
-                    .lessonDate(lessonDate)
-                    .startTime(startTime)
-                    .endTime(startTime.plusMinutes(oneDayClass.getDuration()))
-                    .participantNumber(0)
-                    .oneDayClass(oneDayClass)
-                    .build();
-        }
     }
 
 }
